% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCVfit.R
\name{computeCVfit}
\alias{computeCVfit}
\title{compute fit of cross valudation}
\usage{
computeCVfit(model, ctmodelobj, testset, numObs, model_type = "ctsem",
  fitfunction = "FIML")
}
\description{
compute fit of cross valudation
}
