% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEstimatedParameters.R
\name{getEstimatedParameters}
\alias{getEstimatedParameters}
\title{computes the estimated parameters for a given threshold}
\usage{
getEstimatedParameters(regmodel, zeroThresh = 0.001, setZero = FALSE)
}
\description{
computes the estimated parameters for a given threshold
}
\author{
Jannik Orzek
}
