% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFitMeasures.R
\name{getFitMeasures}
\alias{getFitMeasures}
\title{compute fit measures for laremm}
\usage{
getFitMeasures(regmodel, model_type = "mxModel", fitfun = "FIML",
  ncp_rmsea = FALSE, cvsample = NULL)
}
\description{
compute fit measures for laremm
}
