% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFML.R
\name{getFML}
\alias{getFML}
\title{get FML values}
\usage{
getFML(regmodel, use_unbiasedCov = FALSE)
}
\arguments{
\item{regmodel}{already run regularized model}

\item{use_unbiasedCov}{wheather to use the unbiased covariance; default is FALSE which is equivalent to regsem}
}
\description{
Note: laremm is based on the R package \pkg{regsem}. Because of the early status of laremm, it is recommended to use regsem instead!
Computes FML values corresponding to the values reported in \pkg{regsem} for covariance based models
}
