% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regsem_v_regctsem.R
\name{regsem_v_regctsem}
\alias{regsem_v_regctsem}
\title{compare regsem and laremm}
\usage{
regsem_v_regctsem(mx_regmodel, lav_regmodel, Jacob_equivalent = F)
}
\description{
compare regsem and laremm
}
