% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.CVlaremm.R
\name{summary.CVlaremm}
\alias{summary.CVlaremm}
\title{Summary function for kFoldCV}
\usage{
\method{summary}{CVlaremm}(object, ...)
}
\arguments{
\item{object}{Object from kFoldCV}

\item{...}{additional parameters.}
}
\description{
Summary function for kFoldCV
}
