% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.FitLaremmObject.R
\name{summary.FitLaremmObject}
\alias{summary.FitLaremmObject}
\title{Summary function for FitLaremmObject}
\usage{
\method{summary}{FitLaremmObject}(object, ...)
}
\arguments{
\item{object}{Object from fitRegModels.}

\item{...}{additional parameters.}
}
\description{
Summary function for FitLaremmObject
}
